refactor(visual-diff): split index.ts → overlays + run-diff (SRP)
index.ts (360 lines) mixed CLI parsing, overlay opening, sort-key probe,
diff orchestration, and report writing. Split into three files, each ≤300:

- index.ts (107) — CLI arg parsing + screenshots/diff dispatch only
- overlays.ts (116) — openPrototypeOverlays + openReactOverlays + detectSortKeyMismatch
- run-diff.ts (180) — extract → diff → token hints → write report orchestration

Behavior-preserving: external surface (main(argv)) unchanged. Tests
unchanged (67/67 passing). No new typecheck errors.
